Fix ABI mismatches: seekToSecond f64→f32, AudioBuffer int64→uint64
ma_sound_seek_to_second takes float (f32) in C, but the Zig extern
declared f64, causing the value to be misinterpreted at the FFI boundary.

zaudioAudioBufferConfigInit declared ma_int64 for size_in_frames but
ma_audio_buffer_config_init expects ma_uint64.

+++ src/zaudio.zig
@@ -2765,7 +2765,7 @@ pub const Sound = opaque {
     pub fn seekToSecond(sound: *Sound, seek_point_in_second: f32) Error!void {
         try maybeError(ma_sound_seek_to_second(sound, seek_point_in_second));
     }
-    extern fn ma_sound_seek_to_second(sound: *Sound, seek_point_in_second: f64) Result;
+    extern fn ma_sound_seek_to_second(sound: *Sound, seek_point_in_second: f32) Result;
 
     pub fn getDataFormat(
         sound: *const Sound,
